+++ lib/interactiveCli/register.js
@@ -11,7 +11,7 @@ const {
 } = require('@serverless/platform-sdk');
 const sdkVersion = require('@serverless/platform-sdk/package').version;
 const enableConfirm = require('./enableConfirm');
-const writeOrgAndApp = require('./writeTenantAndApp');
+const writeOrgAndApp = require('./writeOrgAndApp');
 
 const isValidEmail = RegExp.prototype.test.bind(
   new RegExp(
