+++ lib/interactiveCli/set-app.js
@@ -14,7 +14,7 @@ const {
   setDefaultDeploymentProfile,
 } = require('@serverless/platform-sdk');
 const enableConfirm = require('./enableConfirm');
-const writeTenantAndApp = require('./writeTenantAndApp');
+const writeTenantAndApp = require('./writeOrgAndApp');
 const { resolveAccessKey } = require('./utils');
 
 const isValidAppName = RegExp.prototype.test.bind(/^[a-z0-9](?:[a-z0-9-]{0,126}[a-z0-9])?$/);
